Implemented New Standard (Fixed Noah's Mistakes)
