--- src/main/java/frc/auton/AutonOverride.java
@@ -1,5 +1,0 @@
-package frc.auton;
-
-public class AutonOverride {
-    
-}
