--- src/main/java/frc/auton/RobotComponent.java
@@ -1,5 +1,0 @@
-package frc.auton;
-
-public enum RobotComponent {
-    DRIVE, UTIL
-}
