--- src/main/java/frc/auton/mode/AutonMode.java
@@ -1,11 +1,0 @@
-package frc.auton.mode;
-
-/**
- *
- * @author Programmers
- */
-public interface AutonMode {
-
-	public void addToMode(AutonBuilder ab);
-
-}
