+++ src/main/java/frc/io/DriverInput.java
@@ -1,8 +1,8 @@
 package frc.io;
 
-import frc.util.Controller;
-import frc.util.Controller.Axis;
-import frc.util.Controller.Side;
+import frc.util.devices.Controller;
+import frc.util.devices.Controller.Axis;
+import frc.util.devices.Controller.Side;
 
 public class DriverInput {
     private static DriverInput instance;
