--- src/main/java/frc/io/Input/ISensorInput.java
@@ -1,8 +1,0 @@
-package frc.io.Input;
-
-public interface ISensorInput {
-    
-    void reset();
-    void update();
-
-}
