--- src/main/java/frc/io/Output/IRobotOutput.java
@@ -1,7 +1,0 @@
-package frc.io.Output;
-
-public interface IRobotOutput {
-    
-    void stopAll();
-
-}
