+++ src/main/java/frc/io/subsystems/IIO.java
@@ -1,0 +1,9 @@
+package frc.io.subsystems;
+
+public interface IIO {
+    
+    void updateInputs();
+    void resetInputs();
+    void stopAllOutputs();
+
+}
