+++ src/main/java/frc/teleop/TeleopDriver.java
@@ -2,9 +2,9 @@ package frc.teleop;
 
 import frc.io.DriverInput;
 import frc.subsystems.Drive;
-import frc.util.Controller;
-import frc.util.Controller.Axis;
-import frc.util.Controller.Side;
+import frc.util.devices.Controller;
+import frc.util.devices.Controller.Axis;
+import frc.util.devices.Controller.Side;
 
 public class TeleopDriver extends TeleopComponent {
     private static TeleopDriver instance;
